Приступил к заданию - 4.8. Разделяй и властвуй
