+++ src/view/edit-point-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createEventOffer = (offers) => (
   `${offers.length !== 0 ? `<section class="event__section  event__section--offers">
@@ -145,20 +145,13 @@ const createFormEditPointTemplate = (data) => {
     </form>`;
 };
 
-export default class FormEditPointView {
-  #element = null;
+export default class FormEditPointView extends AbstractView {
+
   #routePoint = null;
 
   constructor(routePoint) {
+    super();
     this.#routePoint = routePoint;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
   }
 
   get template() {
