+++ src/view/form-trip-sort-trip-events-view.js
@@ -1,4 +1,4 @@
-import {createElement} from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createFormTripControlsTemplate = () => (
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
@@ -29,22 +29,9 @@ const createFormTripControlsTemplate = () => (
   </form>`
 );
 
-export default class FormTripControlsView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class FormTripControlsView  extends AbstractView {
 
   get template() {
     return createFormTripControlsTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
