+++ src/view/list-empty-view.js
@@ -1,23 +1,10 @@
-import {createElement} from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createListEmptyTemplate = () => '<p class="trip-events__msg">Click New Event to create your first point</p>';
 
-export default class ListEmptyView {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class ListEmptyView extends AbstractView {
 
   get template() {
     return createListEmptyTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
