+++ src/view/trip-events-list-component-view.js
@@ -1,26 +1,13 @@
-import {createElement} from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createTripEventsListComponentTemplate = () => (
   `<li class="trip-events__item">
   </li>`
 );
 
-export default class TripEventsListComponentVeiw {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class TripEventsListComponentVeiw extends AbstractView {
 
   get template() {
     return createTripEventsListComponentTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
