+++ src/view/trip-events-list-view.js
@@ -1,26 +1,13 @@
-import {createElement} from '../utils/render.js';
+import AbstractView from './abstract-view.js';
 
 const createTripEventsListTemplate = () => (
   `<ul class="trip-events__list">
   </ul>`
 );
 
-export default class TripEventsListVeiw {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
+export default class TripEventsListVeiw extends AbstractView {
 
   get template() {
     return createTripEventsListTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
